refactor: separate modules between http, data, infra, utils and config layers
