+++ package.json
@@ -4,10 +4,10 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "tsx watch ./src/server.ts",
+    "dev": "tsx watch ./src/http/server.ts",
     "test": "vitest",
-    "test:unit": "vitest run ./src/**/*.spec.ts",
-    "test:integration": "vitest run ./src/**/*.test.ts",
+    "test:unit": "vitest run ./src/**/**/*.spec.ts",
+    "test:integration": "vitest run ./src/**/**/*.test.ts",
     "test:coverage": "vitest run --coverage",
     "test:ui": "vitest --ui",
     "migrate:dev": "prisma migrate dev",
