+++ src/http/controller/csrf-token-controller.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express'
 
-import { generateToken } from '@/middleware/csrf'
+import { generateToken } from '@/http/middleware/csrf'
 
 export const applyCsrfTokenController = (request: Request, response: Response) => {
   const csrfToken = generateToken(request, response)
