+++ src/http/controller/update-expense-controller.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express'
-import ExpenseRepository from '@/repository/expense-repository'
-import { Expense } from '@/entities/expense'
+import ExpenseRepository from '@/infra/database/repository/expense-repository'
+import { Expense } from '@/data/entities/expense'
 
 class UpdateExpenseController {
   private readonly repository: ExpenseRepository
