+++ src/http/controller/view-expenses-controller.test.ts
@@ -2,10 +2,10 @@ import { afterAll, beforeAll, beforeEach, describe, expect, it } from 'vitest'
 import request from 'supertest'
 import * as falso from '@ngneat/falso'
 
-import app from '@/app'
-import { ExpenseDTO } from '@/dtos/expense'
-import db from '@/database'
-import { MessageErrorDTO } from '@/dtos/error-message'
+import app from '@/http/app'
+import { ExpenseDTO } from '@/data/dtos/expense'
+import db from '@/infra/database'
+import { MessageErrorDTO } from '@/data/dtos/error-message'
 import { generateExpenses } from '@/utils/tests/generate-expenses'
 import { getCSRFTokenAndCookies } from '@/utils/tests/get-csrf-token-and-cookies'
