+++ src/http/factory/add-expense-controller-factory.ts
@@ -1,5 +1,5 @@
-import AddExpenseController from '@/controller/add-expense-controller'
-import ExpenseRepository from '@/repository/expense-repository'
+import AddExpenseController from '@/http/controller/add-expense-controller'
+import ExpenseRepository from '@/infra/database/repository/expense-repository'
 
 export const makeAddExpenseController = () => {
   const expenseRepository = new ExpenseRepository()
