+++ src/http/factory/view-expense-controller-factory.ts
@@ -1,5 +1,5 @@
-import ViewExpensesController from '@/controller/view-expenses-controller'
-import ExpenseRepository from '@/repository/expense-repository'
+import ViewExpensesController from '@/http/controller/view-expenses-controller'
+import ExpenseRepository from '@/infra/database/repository/expense-repository'
 
 export const makeViewExpensesController = () => {
   const expenseRepository = new ExpenseRepository()
