+++ src/http/server.ts
@@ -1,4 +1,4 @@
-import app from '@/app'
+import app from '@/http/app'
 
 app.listen(3333, () => {
   console.log('Server started on port 3333')
