+++ src/infra/database/repository/expense-repository.ts
@@ -1,4 +1,4 @@
-import db from '@/database'
+import db from '@/infra/database'
 
 type AddExpenseRepositoryDTO = {
   description: string
