+++ src/utils/tests/get-csrf-token-and-cookies.ts
@@ -1,5 +1,5 @@
 import request from 'supertest'
-import app from '@/app'
+import app from '@/http/app'
 
 const getCSRFResponseBody = async () => {
   const csrfResponse = await request(app).get('/csrf-token')
